Module aws_sdk_ivs::model

Modules

batch_error

See BatchError

channel

See Channel

channel_summary

See ChannelSummary

destination_configuration

See DestinationConfiguration

playback_key_pair

See PlaybackKeyPair

playback_key_pair_summary

See PlaybackKeyPairSummary

recording_configuration

See RecordingConfiguration

recording_configuration_summary

See RecordingConfigurationSummary

s3_destination_configuration

See S3DestinationConfiguration

stream

See Stream

stream_key

See StreamKey

stream_key_summary

See StreamKeySummary

stream_summary

See StreamSummary

Structs

BatchError

Error related to a specific channel, specified by its ARN.

Channel

Object specifying a channel.

ChannelSummary

Summary information about a channel.

DestinationConfiguration

A complex type that describes a location where recorded videos will be stored. Each member

PlaybackKeyPair

A key pair used to sign and validate a playback authorization token.

PlaybackKeyPairSummary

Summary information about a playback key pair.

RecordingConfiguration

An object representing a configuration to record a channel stream.

RecordingConfigurationSummary

Summary information about a RecordingConfiguration.

S3DestinationConfiguration

A complex type that describes an S3 location where recorded videos will be stored.

Stream

Specifies a live video stream that has been ingested and distributed.

StreamKey

Object specifying a stream key.

StreamKeySummary

Summary information about a stream key.

StreamSummary

Summary information about a stream.

Enums

ChannelLatencyMode

ChannelType

RecordingConfigurationState

StreamHealth

NOTE: StreamHealth::Unknown has been renamed to ::UnknownValue.

StreamState