Module aws_sdk_ivs::operation

Structs

BatchGetChannel

Performs GetChannel on multiple ARNs simultaneously.

BatchGetStreamKey

Performs GetStreamKey on multiple ARNs simultaneously.

CreateChannel

Creates a new channel and an associated stream key to start streaming.

CreateRecordingConfiguration

Creates a new recording configuration, used to enable recording to Amazon S3.

CreateStreamKey

Creates a stream key, used to initiate a stream, for the specified channel ARN.

DeleteChannel

Deletes the specified channel and its associated stream keys.

DeletePlaybackKeyPair

Deletes a specified authorization key pair. This invalidates future viewer tokens

DeleteRecordingConfiguration

Deletes the recording configuration for the specified ARN.

DeleteStreamKey

Deletes the stream key for the specified ARN, so it can no longer be used to

GetChannel

Gets the channel configuration for the specified channel ARN. See also BatchGetChannel.

GetPlaybackKeyPair

Gets a specified playback authorization key pair and returns the arn and

GetRecordingConfiguration

Gets the recording configuration for the specified ARN.

GetStream

Gets information about the active (live) stream on a specified channel.

GetStreamKey

Gets stream-key information for a specified ARN.

ImportPlaybackKeyPair

Imports the public portion of a new key pair and returns its arn and

ListChannels

Gets summary information about all channels in your account, in the AWS region where the

ListPlaybackKeyPairs

Gets summary information about playback key pairs. For more information, see Setting Up Private

ListRecordingConfigurations

Gets summary information about all recording configurations in your account, in the AWS

ListStreamKeys

Gets summary information about stream keys for the specified channel.

ListStreams

Gets summary information about live streams in your account, in the AWS region where the

ListTagsForResource

Gets information about AWS tags for the specified ARN.

PutMetadata

Inserts metadata into the active stream of the specified channel. A maximum of 5 requests

StopStream

Disconnects the incoming RTMPS stream for the specified channel. Can be used in

TagResource

Adds or updates tags for the AWS resource with the specified ARN.

UntagResource

Removes tags from the resource with the specified ARN.

UpdateChannel

Updates a channel's configuration. This does not affect an ongoing stream of this channel.